pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "after-meeting"
version = "0.1.0"
description = "AI-powered meeting audio to minutes/report generator"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "after-meeting contributors" }]
keywords = ["meeting", "minutes", "transcription", "ai", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
]
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-docx>=1.0",
]

[project.optional-dependencies]
stt = ["qwen-asr>=0.0.6", "torch>=2.0"]
pdf = ["docx2pdf>=0.1.8"]
speaker = ["onnxruntime>=1.17", "numpy>=1.26", "scipy>=1.12"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "onnxruntime>=1.17",
    "numpy>=1.26",
    "scipy>=1.12",
]

[project.scripts]
after-meeting = "after_meeting.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/after_meeting"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
